chore: create dedicated component
